# Apply simplify to creation_op and annihilation_op in qubit mapping